task log parser
Implement some log passing for tasks managed to enable recognition for date and time. So for example, if I said doctor's appointment at 6:30, it would recognise there is a date missing and then I will type in the date and then it would recognise and asked me to complete the task by pressing yes. However, we are still facing an issue with the edit button showing and I had to restore back because we couldn't fix it and some files got corrupted so I restored back