Final Tasks and reminders, tweaks
Created a lot of polishing tweaks towards tasking reminders based on the AI and chat interface still need to do a little bit of tweaking on you. I design in the reminders by page pretty much happy with the task page but will be overlooking once the notes page is complete.